Don't clean the archive by default.

+++ clean_support.sh
@@ -1,18 +1,36 @@
 #!/bin/bash
 
-echo "Cleaning archived env setup files..."
-rm -f  archive/*environment_setup*
-echo "Cleaning archived GSL tar balls..."
-rm -rf archive/*gsl*tar.gz
-echo "Cleaning archived HepMC tar balls..."
-rm -rf archive/*HepMC*tar.gz
-# DO NOT remove the log4cpp source bundle we need to install.
-echo "Cleaning archived log4cpp tar balls..." 
-rm -rf archive/1*log4cpp.tgz
-echo "Cleaning archived LHAPDF tar balls..."
-rm -rf archive/*lhapdf*tar.gz
-echo "Cleaning archived Pythia8 tar balls..."
-rm -rf archive/*pythia8*.tgz
+CLEAN_ARCHIVE="no"
+HELPFLAG="no"
+
+while getopts "fh" options; do
+  case $options in
+    f) CLEAN_ARCHIVE="yes";;
+    h) HELPFLAG="yes";;
+  esac
+done
+
+if [ "$HELPFLAG" == "yes" ]; then
+  echo "./clean_support.sh       -f   # clean the user files in the archive directory as well"
+  exit 0
+fi
+
+
+if [ "$CLEAN_ARCHIVE" == "yes" ]; then
+  echo "Cleaning archived env setup files..."
+  rm -f  archive/*environment_setup*
+  echo "Cleaning archived GSL tar balls..."
+  rm -rf archive/*gsl*tar.gz
+  echo "Cleaning archived HepMC tar balls..."
+  rm -rf archive/*HepMC*tar.gz
+  # DO NOT remove the log4cpp source bundle we need to install.
+  echo "Cleaning archived log4cpp tar balls..." 
+  rm -rf archive/1*log4cpp.tgz
+  echo "Cleaning archived LHAPDF tar balls..."
+  rm -rf archive/*lhapdf*tar.gz
+  echo "Cleaning archived Pythia8 tar balls..."
+  rm -rf archive/*pythia8*.tgz
+fi
 
 echo "Cleaning env setup file..."
 rm -f  environment_setup.sh
